Fix a formatting issue with the status bar URL
As one does, one did not notice that the URL added to the status bar
when the show_url setting is turned on did not have a closing ]
character. This is only aesthetic, but it looks wrong for it to not be
there.

Video easter egg: the video that talks about this was recorded prior to
noticing the bug. Phooey.