Use the content type of the URL to set extension
Previously the extension of the file (which would have a direct impact
on the syntax highlighting when the file is opened) always came from the
file being downloaded.

This changes that so that in cases where the filename has no extension,
the mime type is used to set the appropriate extension.